Data.Trie.Internal: reimplemented size with foldl'
The old definition was out of sync with foldl', and so wasn't using
the most efficient of the very many alternative implementations.
We still give a specialized definition rather than calling foldl'
per se and then relying on GHC to inline away the need for a closure.
On our current benchmark suite, this new implementation saves ~25%
time.